requirements.txt
discord.py>=1.7.3
Pillow==9.1.1
PyMySQL>=1.0.2
requests
